Split Svelte fake componet to own file
