+++ examples/svelte/counter/README.md
@@ -1,3 +1,3 @@
 # Counter
 
-A simple counter built with svelte using redux-zero.
+A simple counter built with [Svelte](https://svelte.technology/) using redux-zero.
